fix: move helpers to helper file
